Create findMedianSortedArrays.java
Naive idea: merge/sort and find the median in the merged array, which need O(m + n) time.
To reduce the complexity to O(log(m+n)), we need to eliminate half of the elements at each round.
Key idea: instead of finding the median, we find the k-th smallest element. 
We find the k/2-th smallest element in each array, so that we can eliminate the half that cannot be the required element.

We eliminate k/2 elements at each iteration, when we try to find the median, the complexity is O(log (m+n)).

Note: other methods such as find the median and then eliminate the parts that do not contain the median.
The issue of this method is it is difficult to deal with the odd/even length.